requirements.txt
Ghost.py==0.2.3
PySide==1.2.4
wheel==0.24.0
